Modify tests file names to match corresponding src file names.
